# packages/contracts/openapi/opencore-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/monitor/operations/summary": {
      "get": {
        "operationId": "OperationsController_getSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsSummaryDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Operations"]
      }
    },
    "/api/monitor/jobs": {
      "get": {
        "operationId": "OperationsController_listJobs",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "enabled",
            "required": false,
            "in": "query",
            "schema": { "$ref": "#/components/schemas/Object" }
          },
          {
            "name": "queueName",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDefinitionPageDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Jobs"]
      },
      "post": {
        "operationId": "OperationsController_createJob",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobDefinitionDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Jobs"]
      }
    },
    "/api/monitor/jobs/{code}": {
      "get": {
        "operationId": "OperationsController_getJob",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobDefinitionDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Jobs"]
      },
      "patch": {
        "operationId": "OperationsController_updateJob",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobDefinitionDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Jobs"]
      }
    },
    "/api/monitor/jobs/{code}/enable": {
      "patch": {
        "operationId": "OperationsController_enableJob",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobDefinitionDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Jobs"]
      }
    },
    "/api/monitor/jobs/{code}/disable": {
      "patch": {
        "operationId": "OperationsController_disableJob",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobDefinitionDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Jobs"]
      }
    },
    "/api/monitor/jobs/{code}/trigger": {
      "post": {
        "operationId": "OperationsController_triggerJob",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TriggerJobDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobRunLogDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Jobs"]
      }
    },
    "/api/monitor/jobs/{code}/runs": {
      "get": {
        "operationId": "OperationsController_listJobRuns",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["queued", "running", "completed", "failed"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobRunLogPageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Jobs"]
      }
    },
    "/api/monitor/jobs/{code}/runs/{id}": {
      "get": {
        "operationId": "OperationsController_getJobRun",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobRunLogDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Jobs"]
      }
    },
    "/api/monitor/cache": {
      "get": {
        "operationId": "OperationsController_listCacheKeys",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "prefix",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CacheKeyPageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Cache"]
      }
    },
    "/api/monitor/cache/clear": {
      "post": {
        "operationId": "OperationsController_clearCache",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ClearCacheDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CacheClearResultDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Cache"]
      }
    },
    "/api/monitor/online-users": {
      "get": {
        "operationId": "OperationsController_listOnlineUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": { "$ref": "#/components/schemas/Object" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineUserSessionPageDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Online Users"]
      }
    },
    "/api/monitor/online-users/{id}": {
      "get": {
        "operationId": "OperationsController_getOnlineUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineUserSessionDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Online Users"]
      }
    },
    "/api/monitor/online-users/{id}/kick-out": {
      "post": {
        "operationId": "OperationsController_kickOutSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/KickOutSessionDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineUserSessionDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Monitor Online Users"]
      }
    },
    "/api/optional/reports": {
      "get": {
        "operationId": "OperationsController_listReports",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "enabled",
            "required": false,
            "in": "query",
            "schema": { "$ref": "#/components/schemas/Object" }
          },
          {
            "name": "owner",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDefinitionPageDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Optional Reports"]
      },
      "post": {
        "operationId": "OperationsController_createReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReportDefinitionDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Optional Reports"]
      }
    },
    "/api/optional/reports/{code}": {
      "get": {
        "operationId": "OperationsController_getReport",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReportDefinitionDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Optional Reports"]
      }
    },
    "/api/optional/export-jobs/design": {
      "get": {
        "operationId": "OperationsController_getExportJobDesign",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExportJobDesignDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Operations", "Optional Export Jobs"]
      }
    },
...
  },
  "info": {
    "title": "OpenCore API",
    "description": "OpenCore API contract",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [{ "url": "/" }],
  "components": {
    "schemas": {
...
      "Object": { "type": "object", "properties": {} },
...
      "JobDefinitionSummaryDto": {
        "type": "object",
        "properties": {
          "total": { "type": "number" },
          "enabled": { "type": "number" },
          "disabled": { "type": "number" }
        },
        "required": ["total", "enabled", "disabled"]
      },
      "JobRunSummaryDto": {
        "type": "object",
        "properties": {
          "total": { "type": "number" },
          "queued": { "type": "number" },
          "running": { "type": "number" },
          "completed": { "type": "number" },
          "failed": { "type": "number" },
          "latestStartedAt": { "type": "string" }
        },
        "required": ["total", "queued", "running", "completed", "failed"]
      },
      "CacheSummaryDto": {
        "type": "object",
        "properties": {
          "keyCount": { "type": "number" },
          "totalSizeBytes": { "type": "number" }
        },
        "required": ["keyCount", "totalSizeBytes"]
      },
      "OnlineUserSummaryDto": {
        "type": "object",
        "properties": {
          "total": { "type": "number" },
          "active": { "type": "number" },
          "revoked": { "type": "number" }
        },
        "required": ["total", "active", "revoked"]
      },
      "ReportSummaryDto": {
        "type": "object",
        "properties": {
          "total": { "type": "number" },
          "enabled": { "type": "number" },
          "disabled": { "type": "number" }
        },
        "required": ["total", "enabled", "disabled"]
      },
      "OperationsSummaryDto": {
        "type": "object",
        "properties": {
          "jobs": { "$ref": "#/components/schemas/JobDefinitionSummaryDto" },
          "jobRuns": { "$ref": "#/components/schemas/JobRunSummaryDto" },
          "cache": { "$ref": "#/components/schemas/CacheSummaryDto" },
          "onlineUsers": {
            "$ref": "#/components/schemas/OnlineUserSummaryDto"
          },
          "reports": { "$ref": "#/components/schemas/ReportSummaryDto" },
          "exportJobStatus": { "type": "string" }
        },
        "required": [
          "jobs",
          "jobRuns",
          "cache",
          "onlineUsers",
          "reports",
          "exportJobStatus"
        ]
      },
      "JobDefinitionDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "queueName": { "type": "string" },
          "cron": { "type": "string" },
          "enabled": { "type": "boolean" },
          "retryLimit": { "type": "number" },
          "timeoutSeconds": { "type": "number" },
          "adapter": { "type": "string" },
          "payload": { "type": "object" }
        },
        "required": [
          "id",
          "code",
          "name",
          "queueName",
          "enabled",
          "retryLimit",
          "timeoutSeconds",
          "adapter"
        ]
      },
      "JobDefinitionPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/JobDefinitionDto" }
          },
          "page": { "type": "number" },
          "pageSize": { "type": "number" },
          "total": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"]
      },
      "CreateJobDefinitionDto": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "name": { "type": "string" },
          "queueName": { "type": "string" },
          "cron": { "type": "string" },
          "enabled": { "type": "boolean", "default": true },
          "retryLimit": { "type": "number", "default": 3 },
          "timeoutSeconds": { "type": "number", "default": 60 },
          "payload": { "type": "object" }
        },
        "required": ["code", "name", "queueName"]
      },
      "UpdateJobDefinitionDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "queueName": { "type": "string" },
          "cron": { "type": "string" },
          "enabled": { "type": "boolean" },
          "retryLimit": { "type": "number" },
          "timeoutSeconds": { "type": "number" },
          "payload": { "type": "object" }
        }
      },
      "TriggerJobDto": {
        "type": "object",
        "properties": {
          "actor": { "type": "string" },
          "metadata": { "type": "object" }
        },
        "required": ["actor"]
      },
      "JobRunLogDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "jobCode": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["queued", "running", "completed", "failed"]
          },
          "trigger": { "type": "string", "enum": ["manual", "schedule"] },
          "attempts": { "type": "number" },
          "startedAt": { "type": "string" },
          "finishedAt": { "type": "string" },
          "error": { "type": "string" },
          "metadata": { "type": "object" }
        },
        "required": [
          "id",
          "jobCode",
          "status",
          "trigger",
          "attempts",
          "startedAt"
        ]
      },
      "JobRunLogPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/JobRunLogDto" }
          },
          "page": { "type": "number" },
          "pageSize": { "type": "number" },
          "total": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"]
      },
      "CacheKeyDto": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "prefix": { "type": "string" },
          "ttlSeconds": { "type": "number" },
          "sizeBytes": { "type": "number" }
        },
        "required": ["key", "prefix", "ttlSeconds", "sizeBytes"]
      },
      "CacheKeyPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CacheKeyDto" }
          },
          "page": { "type": "number" },
          "pageSize": { "type": "number" },
          "total": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"]
      },
      "ClearCacheDto": {
        "type": "object",
        "properties": {
          "prefix": { "type": "string" },
          "dryRun": { "type": "boolean", "default": true },
          "confirmed": { "type": "boolean" }
        },
        "required": ["prefix", "dryRun"]
      },
      "CacheClearResultDto": {
        "type": "object",
        "properties": {
          "prefix": { "type": "string" },
          "dryRun": { "type": "boolean" },
          "matchedKeys": { "type": "number" },
          "clearedKeys": { "type": "number" },
          "policy": { "type": "string" }
        },
        "required": ["prefix", "dryRun", "matchedKeys", "clearedKeys", "policy"]
      },
      "OnlineUserSessionDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "username": { "type": "string" },
          "tokenId": { "type": "string" },
          "ip": { "type": "string" },
          "userAgent": { "type": "string" },
          "lastSeenAt": { "type": "string" },
          "expiresAt": { "type": "string" },
          "revokedAt": { "type": "string" },
          "revokedBy": { "type": "string" },
          "revokedReason": { "type": "string" }
        },
        "required": [
          "id",
          "username",
          "tokenId",
          "ip",
          "userAgent",
          "lastSeenAt",
          "expiresAt"
        ]
      },
      "OnlineUserSessionPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OnlineUserSessionDto" }
          },
          "page": { "type": "number" },
          "pageSize": { "type": "number" },
          "total": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"]
      },
      "KickOutSessionDto": {
        "type": "object",
        "properties": {
          "actor": { "type": "string" },
          "reason": { "type": "string" }
        },
        "required": ["actor", "reason"]
      },
      "ReportDefinitionDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "querySchema": { "type": "object" },
          "enabled": { "type": "boolean" },
          "owner": { "type": "string" }
        },
        "required": ["id", "code", "name", "querySchema", "enabled", "owner"]
      },
      "ReportDefinitionPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReportDefinitionDto" }
          },
          "page": { "type": "number" },
          "pageSize": { "type": "number" },
          "total": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"]
      },
      "CreateReportDefinitionDto": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "querySchema": { "type": "object" },
          "enabled": { "type": "boolean", "default": true },
          "owner": { "type": "string" }
        },
        "required": ["code", "name", "querySchema", "owner"]
      },
      "ExportJobDesignDto": {
        "type": "object",
        "properties": {
          "resource": { "type": "string" },
          "status": { "type": "string" },
          "requiredBindings": {
            "type": "array",
            "items": { "type": "string" }
          },
          "safetyChecks": { "type": "array", "items": { "type": "string" } },
          "runbook": { "type": "string" }
        },
        "required": [
          "resource",
          "status",
          "requiredBindings",
          "safetyChecks",
          "runbook"
        ]
      },
...
    }
  }
}